Ask the server not to send metadata in the response to prepared statements
In v2 prepared statements get metadata about their results along with the metadata about the parameters. When a prepared statement has this data it can ask the server not to send it on every response, hopefully saving a lot of bandwidth.

This patch introduces LazyQueryResult and RawRowsResultResponse that make it possible to defer row decoding to the absolute last moment. It also adds PreparedStatement#result_metadata for access to the result metadata.

All decoder methods also get the frame length, even though it's only RowsResultResponse that uses it.